requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-ckeditor==6.0.0
django-js-asset==1.2.2
Pillow==8.2.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
